fix window change misalignment
I fixed the issue of the board getting misaligned when window size changes. Instead of saving the initial position of the block div, I keep a reference to the board div and make position calculations using delta(boardDiv, dropTargetDiv). why? "a `ClientRect` object will reflect the element's dimensions and position **after the transform has been applied**." Simply we need an object (like board) that stays in the same place.